Rename Tinybox file to lowercase
